dept: Fix a missing oppertunities for more detection by irq
The original code propagates iwait when a new iwait is intoduced or when
iwait == DEPT_IWAIT_UNKNOWN is detected.

However, iwait should also be propagated when a new dependency is added
into the dependency graph but the original code didn't. So made it
propagate properly when a new dependency is added as expected.

Signed-off-by: Byungchul Park <byungchul.park@lge.com>