dept: Work with only real sleeps when committing a wait in __schedule()
Dept commits a staged wait in __schedule() if it exists. On the commit,
Dept considers it a potential sleep, even if the sleep has been woken up
before getting to __schedule(), in order to work with potential ones.

This helps Dept do stronger detection but leads false positives. Let
Dept work only with real sleeps conservatively for now. So did it.

Signed-off-by: Byungchul Park <byungchul.park@lge.com>