dept: fix reference of invalidated dept_class
Some dept_class could be refered after the key where class belong to are
destroy if another key caches the class.
This wrong reference could make unexcepected problem (i.e) boot failure.

To address this, caching the class when the key is local only
when class is created and refer it dept_lock section.

Signed-off-by: Yeoreum Yun <yeoreum.yun@arm.com>